+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 0.12.0
+
 - Only add to browser history when updating non-location settings
 - Add support for Leaflet styles
 - Document gazetteer configuration